% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.AT}
\alias{f.AT}
\title{Coupled conductance photosynthesis model and energy balance model}
\usage{
f.AT(PFD, cs, Tair, RH, wind, precision = 0.1, max_it = 10, param)
}
\arguments{
\item{wind}{Wind speed at the surface of the leaf in m.s-1}

\item{precision}{Precision of the leaf temperature prediction. The resolution of the energy balance coupled with the photosynthesis and stomatal conductance is numerical. The smaller the precision, the longer will be the resolution.}

\item{max_it}{Maximum number of iterations to find the solution}

\item{param}{List of parameters given by f.make.param()}
}
\value{
- A: Raw assimilation of the leaf in micromol.m-2.s-1
 - Ag: Gross assimilation in micromol.m-2.s-1
 - gs: Conductance of the leaf for water vapour in mol m-2 s-1
 - ci: Intracellular CO2 concentration in micromol.mol-1
 - cc: Mesophyll CO2 concentration in micromol.mol-1 (for the models using mesophyll conductance)
 - ds: Leaf surface to air vapour pressure deficit in Pa
 - Trans: Water transpiration in mL m-2 s-1
 - Tleaf: Leaf Temperature in K
}
\description{
Coupled conductance photosynthesis model and energy balance model
}
\details{
This function allows to calculate the photosynthesis from environmental variables PFD, RH, wind, cs and Tair.
The energy balance model is calculated using the package Tealeaves (see reference). The energy balance calculation involves the stomatal conductance and the cuticular conductance.
Here the cuticular conductance is considered to be equal to g0 as done in some TBMs even if it is probably a wrong representation.This choice was made to prevent unrealistic energy budgets when the conductance is too low (<= 0) for low light levels.
}
\examples{
f.AT(PFD=1500,cs=400,Tair=299,wind=2,RH=70,param=f.make.param())
}
\references{
tealeaves: an R package for modelling leaf temperature using energy budgets. Christopher. D. Muir. bioRxiv 529487; doi: https://doi.org/10.1101/529487
}
