% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.ATnotvectorised}
\alias{f.ATnotvectorised}
\title{Coupled conductance photosynthesis model and energy balance model}
\usage{
f.ATnotvectorised(
  PFD,
  ca,
  Tair,
  RHa,
  wind,
  precision = 0.05,
  max_it = 10,
  param,
  NIR = NA,
  abso_s = 0.5,
  leaf_size = 0.04
)
}
\arguments{
\item{PFD}{Photosynthetic light at the leaf surface in micro mol m-2 s-1}

\item{ca}{CO2 concentration in the air}

\item{Tair}{Air temperature in Kelvin}

\item{RHa}{Relative humidity in the air (between 0 and 100)}

\item{wind}{Wind speed at the surface of the leaf in m.s-1}

\item{precision}{Precision of the leaf temperature prediction. The resolution of the energy balance coupled with the photosynthesis and stomatal conductance is numerical. The smaller the precision, the longer will be the resolution.}

\item{max_it}{Maximum number of iterations to find the solution}

\item{param}{List of parameters given by f.make.param()}

\item{NIR}{NIR radiation in watt m-2, if not given, then by default the shortwave radiation is calculated as PFD/4.57+NIR = PFD/(4.57*0.45) (see Yun et al. 2020 for the constants 0.45 and 4.57 )}

\item{abso_s}{absorptance of the leaves in shortwave, see documentation of tealeaves package}

\item{leaf_size}{Dimension of the leaf in m (see Tealeaves package)}
}
\value{
- A: Raw assimilation of the leaf in micromol.m-2.s-1
 - Ag: Gross assimilation in micromol.m-2.s-1
 - gs: Conductance of the leaf for water vapour in mol m-2 s-1
 - ci: Intracellular CO2 concentration in micromol.mol-1
 - cc: Mesophyll CO2 concentration in micromol.mol-1 (for the models using mesophyll conductance)
 - ds: Leaf surface to air vapour pressure deficit in Pa
 - Trans: Water transpiration in mL m-2 s-1
 - Tleaf: Leaf Temperature in K
}
\description{
Coupled conductance photosynthesis model and energy balance model
}
\details{
This function allows to calculate the photosynthesis from environmental variables PFD, RH, wind, cs and Tair.
The energy balance model is calculated using the package Tealeaves (see reference). The energy balance calculation involves the stomatal conductance and the cuticular conductance.
Here the minimum conductance is considered to be equal to g0.
}
\examples{
f.ATnotvectorised(PFD=1500,ca=400,Tair=298,wind=2,RHa=70,param=f.make.param(g0=0.03))
}
\references{
tealeaves: an R package for modelling leaf temperature using energy budgets. Christopher. D. Muir. bioRxiv 529487; doi: https://doi.org/10.1101/529487 
Yun, S. H., Park, C. Y., Kim, E. S., & Lee, D. K. (2020). A Multi-Layer Model for Transpiration of Urban Trees Considering Vertical Structure. Forests, 11(11), 1164.
}
\keyword{internal}
