% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.AWc}
\alias{f.AWc}
\title{Carbon assimilation under rubisco carboxylation limitation}
\usage{
f.AWc(
  ci,
  Tleaf,
  param = list(ci, Tleaf, RdHa, GstarRef, GstarHa, RdRef, VcmaxRef, VcmaxHa, VcmaxHd,
    VcmaxS, KoRef, KoHa, KcRef, KcHa)
)
}
\arguments{
\item{Tleaf}{Temperature of the leaf in Kelvin}

\item{param}{List of parameters, see f.make.param for details}
}
\value{

}
\description{
Carbon assimilation under rubisco carboxylation limitation
}
\examples{
plot(x=seq(1,1500,10),y=f.AWc(ci=seq(1,1500,10),Tleaf=300,param=f.make.param()),xlab='Intracellular CO2 in ppm',ylab='Awc in micromol.m-2.s-1')
}
