% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.AWj}
\alias{f.AWj}
\title{Carbon assimilation under electron transport limitation}
\usage{
f.AWj(
  PFD,
  ci,
  Tleaf,
  param = list(RdHa, GstarRef, GstarHa, JmaxHa, JmaxHd, JmaxS, JmaxRef, RdRef, abso, f,
    LogitTheta)
)
}
\arguments{
\item{PFD}{Photo Flux Density in micromol.m-2,s-1}

\item{Tleaf}{Temperature of the leaf in Kelvin}

\item{param}{List of parameters, see f.make.param for details}
}
\value{
Assimilation under Jmax limitation
}
\description{
Carbon assimilation under electron transport limitation
}
\examples{
plot(x=seq(0,1500,10),y=f.AWj(PFD=seq(0,1500,10),ci=270,Tleaf=300,param=f.make.param()),xlab='PFD',ylab='Awj in micromol.m-2.s-1')
}
