% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.CO2.fitting}
\alias{f.CO2.fitting}
\title{Fitting function for Aci data}
\usage{
f.CO2.fitting(
  measures,
  id.name = NULL,
  Start = list(JmaxRef = 90, VcmaxRef = 70, RdRef = 1),
  param = f.make.param(),
  modify.init = TRUE,
  do.plot = TRUE
)
}
\arguments{
\item{measures}{Data frame of measures obtained from gas exchange analyser with at least the columns Photo, Ci, PARi and Tleaf (in K)}

\item{id.name}{Name of the colums in measures with the identifier for the curve.}

\item{Start}{List of parameters to fit with their initial values.}

\item{param}{See f.make.param() for details.}

\item{modify.init}{TRUE or FALSE, allows to modify the Start values before fitting the data}

\item{do.plot}{TRUE or FALSE, plot data and fitted curves ?}
}
\value{

}
\description{
Function to fit f.Aci model to data. The parameters to fit have to be described in the list Start.
All the other parameters of the f.Aci functions have to be in param. If the parameters from Start are repeated in param, the later one will be ignored.
This function uses two methods to fit the data. First by minimizing the residual sum-of-squares of the residuals and then by maximizing the likelihood function. The first method is more robust but the second one allows to calculate the confident interval of the parameters.
}
\examples{
##Simulation of a CO2 curve
data=data.frame(Tleaf=rep(300,20),
Ci=seq(40,1500,75),PARi=rep(2000,20),Photo=f.Aci(PFD=2000,Tleaf=300,ci=seq(40,1500,75),
param=f.make.param(RdRef=1.25,VcmaxRef=57,JmaxRef=92))+rnorm(n = 20,mean = 0,sd = 0.5))

f.CO2.fitting(measures=data,id.name=NULL,Start=list(JmaxRef=90,VcmaxRef=70,RdRef=1))
}
