% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canopy_scaling.R
\name{f.GPP}
\alias{f.GPP}
\title{Canopy scale GPP calculation}
\usage{
f.GPP(
  meteo_hourly,
  Vcmax_Profile,
  Jmax_Profile,
  Rd_Profile,
  Tp_Profile,
  g0_Profile,
  g1_Profile,
  gsmin,
  canopy,
  Patm = 100,
  ...
)
}
\arguments{
\item{meteo_hourly}{See f.canopy.interception doc. In addition to the requirement for f.canopy.interception, the leaf temperature has to be informed within the column Tleaf.}

\item{Vcmax_Profile}{Vector of the values of Vcmax at the reference temperature at each layer of the canopy.}

\item{Jmax_Profile}{Vector of the values of Jmax at the reference temperature at each layer of the canopy.}

\item{Rd_Profile}{Vector of the values of Rd at the reference temperature at each layer of the canopy.}

\item{Tp_Profile}{Vector of the values of Tp at the reference temperature at each layer of the canopy.}

\item{g0_Profile}{Vector of the values of g0 at the reference temperature at each layer of the canopy.}

\item{g1_Profile}{Vector of the values of g1 at the reference temperature at each layer of the canopy.}

\item{gsmin}{Minimum stomatal conductance for water to consider. This value will be used as the minimum conductance value to avoid 0 and negative values obtained from the coupled assimilation and conductance models.}

\item{canopy}{Description of the canopy interception (see canopy_interception function).}

\item{Patm}{Atmospheric pressure (used to calculate the transpiration).}

\item{...}{Other parameters of the photosynthetic model, without gradients, for example curvature factor, quantum yield.. see the help of f.make.param().}
}
\description{
Generic function to calculate the GPP within a forest (Here GPP = sum of Anet at the canopy level, so it takes into account the leaf mitochondrial respiration)
}
\examples{
#See vignettes on github
}
