% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canopy_scaling.R
\name{f.GPPT}
\alias{f.GPPT}
\title{Canopy scale GPP calculation, with leaf energy budget}
\usage{
f.GPPT(
  TBM,
  meteo_hourly,
  Vcmax_Profile,
  Jmax_Profile,
  Rd_Profile,
  Tp_Profile,
  g0_Profile,
  g1_Profile,
  gsmin,
  canopy,
  Patm = 100,
  ...
)
}
\arguments{
\item{TBM}{Specific TBM to use (ORCHIDEE, CLM4.5, FATES or JULES)}

\item{meteo_hourly}{Hourly weather data frame with at least the column at (air temperature in degree C) RH (humidity in pc) and sr the total PAR in micro mol m-2 s-1}

\item{Vcmax_Profile}{Vector of the values of Vcmax at the reference temperature at each layer of the canopy}

\item{Jmax_Profile}{Vector of the values of Jmax at the reference temperature at each layer of the canopy}

\item{Rd_Profile}{Vector of the values of Rd at the reference temperature at each layer of the canopy}

\item{Tp_Profile}{Vector of the values of Tp at the reference temperature at each layer of the canopy}

\item{g0_Profile}{Vector of the values of g0 at the reference temperature at each layer of the canopy}

\item{g1_Profile}{Vector of the values of g1 at the reference temperature at each layer of the canopy}

\item{gsmin}{Minimum stomatal conductance for water to consider. This value will be used as the minimum conductance value to avoid 0 and negative values obtained from the coupled assimilation and conductance models}

\item{canopy}{Description of the canopy interception (see canopy_interception function)}

\item{Patm}{Atmospheric pressure (used to calculate the transpiration)}

\item{...}{Other parameters of the photosynthetic model, without gradients, for example curvature factor, quantum yield.. see the help of f.make.param()}
}
\value{

}
\description{
Generic function to calculate the GPP within a forest (Here GPP = sum of Anet at the canopy level, so it takes into account the leaf mitochondrial respiration)
}
\examples{
## Simulation of photosynthetic gradients
LAI=seq(0,6,6/14)
dLAI=rep(6/15,15)
Vcmax=f.VcmaxRef.LAI(kn=0.11,LAI=LAI,Vcmax0=70)
Jmax=1.7*Vcmax; Tp=1/5*Vcmax; Rd=0.03*Vcmax
##Simulation of weather data
meteo_hourly=data.frame(time=0:23,RH=80,Tair=25,PFD=sin(seq(0,pi,pi/23))*2000,Tleaf=25,wind=2,NIR=sin(seq(0,pi,pi/23))*2000/4.57)
meteo_hourly[!meteo_hourly$time\%in\%7:17,'PFD']=0
meteo_hourly[!meteo_hourly$time\%in\%7:17,'NIR']=0
##Representation of the light interception inside the canopy
canopy=f.canopy.interception(meteo_hourly=meteo_hourly,lat = 9.2801048,t.d = 0:23,DOY = 60,nlayers = 15,dLAI = dLAI)
GPP_sc1=f.GPPT(TBM = "FATES",meteo_hourly = meteo_hourly,Vcmax_Profile = Vcmax,
Jmax_Profile =Jmax ,Rd_Profile =Rd ,Tp_Profile = Tp,
g0_Profile = rep(0.02,length(Vcmax)),g1_Profile = rep(4,length(Vcmax)),canopy=canopy,gsmin = 0.01)
}
