% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{f.import_licor6400}
\alias{f.import_licor6400}
\title{Import Licor 6400 file}
\usage{
f.import_licor6400(
  file,
  column_display = c("Photo", "Cond", "PARi", "Ci", "Leaf_Barcode", "Species",
    "Tree Canopy", "Age", "file")
)
}
\arguments{
\item{file}{File to import by the function}

\item{column_display}{The first lines of the file which are part of this list are displayed by this function after being imported.}
}
\value{
dataframe
}
\description{
This functions allows to import the text file produced by LICOR as a data.frame
}
\references{
Adapted from http://www.ericrscott.com/2018/01/17/li-cor-wrangling/
}
