% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.light.fitting}
\alias{f.light.fitting}
\title{Fitting function for AQ data}
\usage{
f.light.fitting(
  measures,
  id.name = NULL,
  Start = list(JmaxRef = 90, LogitTheta = 0.6, RdRef = 1),
  param = f.make.param(),
  modify.init = TRUE
)
}
\arguments{
\item{measures}{Data frame of measures obtained from gas exchange analyser with at least the columns Photo, Ci, PARi and Tleaf (in K)}

\item{id.name}{Name of the colums in measures with the identifier for the curve.}

\item{Start}{List of parameters to fit with their initial values.}

\item{param}{See f.make.param() for details.}

\item{modify.init}{TRUE or FALSE, allows to modify the Start values before fitting the data}
}
\value{

}
\description{
Fitting function for AQ data
}
\examples{
data=data.frame(Tleaf=300,Ci=280,PARi=seq(0,2000,67),Photo=f.AWj(PFD=seq(0,2000,67),Tleaf=300,ci=280,param=f.make.param(LogitTheta=f.logit(0.89),JmaxRef=215,RdRef=1))+rnorm(n = 30,mean = 0,sd = 0.35))
Start=list(JmaxRef=50,LogitTheta=f.logit(0.8),RdRef=0.5)
f.light.fitting(measures=data,Start=Start)
}
