% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.logistic}
\alias{f.logistic}
\title{Logistic function}
\usage{
f.logistic(x)
}
\value{

}
\description{
This function takes it values in -Inf;+Inf and returns values in 0;1. It is the inverse function of f.logit, ie f.logistic(logit(x))=x
}
\details{
f.logistic(x)=1/(1+exp(-x)) if x<0, = exp(x)/(1+exp(x)) if x<=0
}
\examples{
plot(x=seq(-10,10,0.1),y=f.logistic(x=seq(-10,10,0.1)))
}
