% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\encoding{UTF-8}
\name{f.logit}
\alias{f.logit}
\title{Function logit}
\usage{
f.logit(x)
}
\value{

}
\description{
This function takes it values in 0;1- and returns values in Inf;+Inf. It is the inverse function of f.logistic
}
\examples{
plot(x=seq(0,1,0.01),y=f.logit(x=seq(0,1,0.01)))
}
