% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.MinusLogL_Aq}
\alias{f.MinusLogL_Aq}
\title{Title}
\usage{
f.MinusLogL_Aq(
  data,
  sigma,
  R = 8.314,
  O2 = 210,
  TRef = 298.16,
  Patm = 101,
  JmaxRef = 85,
  JmaxHa = 50300,
  JmaxHd = 152044,
  JmaxS = 495,
  VcmaxRef = 55,
  VcmaxHa = 73637,
  VcmaxHd = 149252,
  VcmaxS = 486,
  RdRef = 1,
  RdHa = 46390,
  KcRef = 404.9,
  KcHa = 79430,
  KoRef = 278.4,
  KoHa = 36380,
  GstarRef = 42.75,
  GstarHa = 37830,
  abso = 0.85,
  f = 0.15,
  LogitTheta = f.logit(0.85),
  g0 = 0.01,
  g1 = 2
)
}
\arguments{
\item{sigma}{Sigma value}

\item{R}{Ideal gas constant}

\item{O2}{O2 concentration in ppm}

\item{TRef}{Reference temperature for Kc, Ko, Rd,GammaStar Vcmax, Jmax}

\item{Patm}{Atmospheric pressure in Pa}

\item{JmaxRef}{Maximum electron transport rate in micromol.m-2.s-1}

\item{JmaxHa}{Energy of activation for Jmax in J.mol-1}

\item{JmaxHd}{Energy of desactivation for Jmax in J.mol-1}

\item{JmaxS}{Entropy term for Jmax in J.mol-1.K-1}

\item{VcmaxRef}{Maximum rate of Rubisco for carboxylation micromol.m-2.s-1}

\item{VcmaxHa}{Energy of activation for Vcmax in J.mol-1}

\item{VcmaxHd}{Energy of desactivation for Vcmax in J.mol-1}

\item{VcmaxS}{Entropy term for Vcmax in J.mol-1.K-1}

\item{RdRef}{Respiration value at the reference temperature}

\item{RdHa}{Energie of activation for Rd in J.mol-1}

\item{KcRef}{Michaelis-Menten constant of Rubisco for CO2 at the reference temperature in micromol.mol-1}

\item{KcHa}{Energy of activation for Kc in J.mol-1}

\item{KoRef}{ichaelis-Menten constant of Rubisco for CO2 at the reference temperature in milimol.mol-1}

\item{KoHa}{Energy of activation for Ko in J.mol-1}

\item{GstarRef}{CO2 compensation point in absence of respiration in micromol.mol-1}

\item{GstarHa}{Enthalpie of activation for Gstar in J.mol-1}

\item{abso}{Absorptance of the leaf in the photosynthetic active radiation wavelenghts}

\item{f}{Correcting factor for the spectral quality of the light}

\item{LogitTheta}{Theta is the empirical curvacture factor for the response of J to PFD. It takes its values between 0 and 1. To avoid numerical issues when fitting data, this parameters is transformed in this model and called LogitTheta. The transformation between Theta and LogitTheta is: Theta=f.logistic(LogitTheta) and LogitTheta=f.logit(Theta)}

\item{g0}{Constant of the USO model, representing the conductance when A is 0, in mol.m-2.s-1}

\item{g1}{Slope parameter, between 1.14 and 3.58 KPa^0.5 (Wu et al., 2019)}
}
\value{

}
\description{
Title
}
\keyword{internal}
