% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.modified.arrhenius.inv}
\alias{f.modified.arrhenius.inv}
\title{Temperature dependence of Jmax and Vcmax}
\usage{
f.modified.arrhenius.inv(P, Ha, Hd, s, Tleaf, TRef = 298.16, R = 8.314)
}
\arguments{
\item{P}{Value of the parameter, here Vcmax or Jmax, at the leaf temperature in micromol.m-2.s-1}

\item{Ha}{Energy of activation in J.mol-1}

\item{Hd}{Energy of desactivation in J.mol-1}

\item{s}{Entropy term in J.mol-1.K-1}

\item{Tleaf}{Temperature of the leaf in Kelvin}

\item{TRef}{Reference temperature}

\item{R}{Ideal gas constant}
}
\value{

}
\description{
Retrieve the reference temperature value of a parameter knowing its value at Tleaf
}
