% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.plot.Aci}
\alias{f.plot.Aci}
\title{Plot Aci data and model}
\usage{
f.plot.Aci(measures = NULL, list_legend, param, name = "")
}
\arguments{
\item{measures}{Data frame obtained from CO2 curve with at least columns Photo, Ci, PARi and Tleaf}

\item{param}{List of parameters, see f.make.param for details}

\item{name}{Name of the curve to be displayed}
}
\value{
Plot a figure
}
\description{
Plot a generic graphic with observed data and predictions. Be careful to sort the data.frame beforehand.
}
\examples{
param=f.make.param()
Photo=f.Aci(PFD=2000,Tleaf=300,ci=seq(40,1500,50),param=param)+rnorm(n = 30,mean = 0,sd = 0.5)
data=data.frame(Tleaf=rep(300,30),Ci=seq(40,1500,50),PARi=rep(2000,30),Photo=Photo)
f.plot.Aci(measures=data,param=param,list_legend=param['VcmaxRef'],name='Example 01')
}
