% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.plot.AQ}
\alias{f.plot.AQ}
\title{Plot AQ data and model}
\usage{
f.plot.AQ(measures = NULL, param, list_legend = NULL, name = "")
}
\arguments{
\item{measures}{Data frame obtained from CO2 curve with at least columns Photo, Ci, PARi and Tleaf}

\item{param}{List of parameters, see f.make.param for details}

\item{name}{Name of the curve to be displayed}
}
\value{
Plot a figure
}
\description{
Plot a generic graphic with observed data and predictions. Be careful to sort the data.frame beforehand.
}
\examples{
param=f.make.param()
data=data.frame(Tleaf=seq(298,305,0.24),Ci=seq(300,271),PARi=seq(0,2000,67),Photo=f.AWj(PFD=seq(0,2000,67),Tleaf=seq(298,305,0.24),ci=seq(300,271),param=param)+rnorm(n = 30,mean = 0,sd = 0.5))
f.plot.AQ(measures=data,param=param,list_legend=param[c('VcmaxRef','RdRef','JmaxRef')],name='Example 01')
}
