% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.Q10.modified}
\alias{f.Q10.modified}
\title{Temperature dependence of photosynthetic parameters}
\usage{
f.Q10.modified(Pref, Q10, Tleaf, TRef, Tlow, Tup)
}
\arguments{
\item{TRef}{Reference temperature for Kc, Ko, Rd,GammaStar Vcmax, Jmax}
}
\value{
Value of the photosynthetic parameter at the specified leaf temperature
}
\description{
Temperature dependence of photosynthetic parameters
}
\details{
This equation is used in JULES TBM model
}
\references{
Clark, D. B., Mercado, L. M., Sitch, S., Jones, C. D., Gedney, N., Best, M. J., . Cox, P. M. (2011). The Joint UK Land Environment Simulator (JULES), model description - Part 2: Carbon fluxes and vegetation dynamics. Geoscientific Model Development, 4(3), 701-722. doi:10.5194/gmd-4-701-2011
}
