% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.smooth}
\alias{f.smooth}
\title{Smoothing functions between photosynthesis limitations (for example between rubisco carboxylation and light limitation)}
\usage{
f.smooth(A1, A2, theta)
}
\arguments{
\item{theta}{Smoothing factor}
}
\value{
Smoothed value
}
\description{
Smoothing functions between photosynthesis limitations (for example between rubisco carboxylation and light limitation)
}
\examples{
A1= seq(0,20,1)
A2= seq(9,11,2/20)
Asmooth=f.smooth(A1=A1,A2=A2,theta=0.99)
plot(A1,type='l')
lines(A2)
lines(Asmooth,col='blue')
}
