% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.solv.Acc}
\alias{f.solv.Acc}
\title{Analytical solution of the coupled photosynthesis with mesophyll conductance and USO model}
\usage{
f.solv.Acc(x1, x2, g0, g1, power, ds, RH, model, gm, Rd, Gstar, cs)
}
\arguments{
\item{cs}{}
}
\value{

}
\description{
Analytical solution of the coupled photosynthesis with mesophyll conductance and USO model
}
\keyword{internal}
