% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.solv.cc}
\alias{f.solv.cc}
\title{Analytical solution for f.Aci with mesophyll conductance}
\usage{
f.solv.cc(x1, x2, gm, Rd, Gstar, ci)
}
\arguments{
\item{ci}{}
}
\value{

}
\description{
Analytical solution for f.Aci with mesophyll conductance
}
\keyword{internal}
