% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.solv}
\alias{f.solv}
\title{Analytical solution of the coupled photosynthesis and USO model}
\usage{
f.solv(x, y, z, cs, Rd, Gstar, g0, g1, power, ds)
}
\arguments{
\item{ds}{}
}
\value{

}
\description{
Analytical solution of the coupled photosynthesis and USO model
}
\keyword{internal}
