% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.SumSq_Aq}
\alias{f.SumSq_Aq}
\title{Compute the sum square of the difference between obervations and predictions}
\usage{
f.SumSq_Aq(Fixed, data, Start)
}
\arguments{
\item{data}{Data frame obtained from CO2 curve with at least columns Photo, Ci, PARi and Tleaf}

\item{x}{List of parameters to fit}
}
\value{
Sum square of the difference between predictions and observations
}
\description{
Function used to fit the parameters of a CO2 curve
}
\keyword{internal}
