% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.A}
\alias{f.A}
\title{Coupled conductance photosynthesis model}
\usage{
f.A(PFD, cs, Tleaf, Tair, RH, param = f.make.param())
}
\value{
List of different variables:
 - A: Raw assimilation of the leaf in micromol.m-2.s-1
 - gs: Conductance of the leaf for water vapour
 - ci: Intracellular CO2 concentration in micromol.mol-1
 - cc: Mesophyll CO2 concentration in micromol.mol-1 (for the models using mesophyll conductance)
 - ds: Leaf surface to air vapour pressure deficit in Pa
}
\description{
Photosynthesis model at the leaf level using the farquhar equations. The parameters can be defined by the
function f.make param and corresponds to the parameters inplemented in different Terrestrial Biosphere Modesl such as ORCHIDEE, JULES, CLM4.5 or FATES
}
\examples{
f.A(PFD=2000,cs=400,Tleaf=273.16+29,Tair=273.16+28,RH=70,param=f.make.param())
}
