% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.Aci}
\alias{f.Aci}
\title{Photosynthesis model}
\usage{
f.Aci(PFD, ci, Tleaf, param = f.make.param())
}
\arguments{
\item{PFD}{Photosynthetic light at the leaf surface in micro mol m-2 s-1.}

\item{ci}{Intercellular CO2 concentration in ppm}

\item{Tleaf}{Leaf temperature in Kelvin.}

\item{param}{List of parameters, see f.make.param for details}
}
\value{
List of different variables:
 - A: Raw assimilation of the leaf in micromol.m-2.s-1.
 - Ac: Rubisco limitation assimilation of the leaf in micromol.m-2.s-1.
 - Aj: Electron transport rate assimilation of the leaf in micromol.m-2.s-1.
 - Ap: TPU rate of the leaf in micromol.m-2.s-1.
 - Ag: Gross assimilation in micromol.m-2.s-1.
}
\description{
Calculate the assimilation according to Farquhar equations. Contrary to f.A, this function uses inter-cellular CO2 and not ambient air CO2
}
\examples{
ci=seq(40,1500,10)
plot(x=ci,y=f.Aci(PFD=2000,ci=ci,Tleaf=300,param=f.make.param())$A)
}
