% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.Aci}
\alias{f.Aci}
\title{Photosynthesis model}
\usage{
f.Aci(PFD, ci, Tleaf, param = list())
}
\arguments{
\item{param}{List of parameters, see f.make.param for details}
}
\value{
Assimilation in micromol.m-2.s-1
}
\description{
Calculate the assimilation according to Farquhar equations. Contrary to f.A, this function uses intracellular CO2 and not ambiant air CO2
}
\examples{
ci=seq(40,1500,10)
plot(x=ci,y=f.Aci(PFD=2000,ci=ci,Tleaf=300,param=f.make.param())$A)
}
