% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.arrhenius.inv}
\alias{f.arrhenius.inv}
\title{Temperature dependence of Gamma star, Ko, Kc and Rd}
\usage{
f.arrhenius.inv(P, Ha, Tleaf, TRef = 298.16, R = 8.314)
}
\arguments{
\item{P}{Value of the parameter at Tleaf}

\item{Ha}{Enthalpie of activation in J.mol-1}

\item{Tleaf}{Temperature of the leaf in Kelvin}

\item{TRef}{Reference temperature}

\item{R}{Ideal gas constant}
}
\value{

}
\description{
Temperature dependence of Gamma star, Ko, Kc and Rd
}
\details{
Retrieve the value of the parameter at Tref knowing its value at Tleaf
}
