% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.AT}
\alias{f.AT}
\title{Coupled conductance photosynthesis model with energy balance model}
\usage{
f.AT(
  PFD,
  cs,
  Tair,
  RH,
  wind,
  precision = 0.1,
  max_it = 10,
  param = f.make.param()
)
}
\arguments{
\item{wind}{Wind speed at the surface of the leaf in m.s-1}

\item{precision}{Precision of the leaf temperature prediction. The resolution of the energy balance coupled with the photosynthesis and stomatal conductance is numerical. The smaller the precision, the longer will be the resolution.}

\item{max_it}{Maximum number of iterations to find the solution}

\item{param}{List of parameters given by f.make.param()}
}
\value{

}
\description{
Coupled conductance photosynthesis model with energy balance model
}
\details{
This function allo to calculate the photosynthesis from environmental variables PFD, RH, wind, cs, Tair.
(There is no boundary layer). The energy balance model comes from the package Tealeaves (see reference). The energy balance calculation involves the stomatal conductance and the cuticular conductance.
Here the cuticular conductance is considered to be equal to g0 even if it is wrong for the USO models. Most of the times, no precaution is taken on gs_min when fitting the conductance models so gs_min is often negative. This choice was made to prevent unrealistic energy budgets.
}
\examples{
f.AT(PFD=1500,cs=400,Tair=299,wind=2,RH=70,param=f.make.param())
}
\references{
tealeaves: an R package for modelling leaf temperature using energy budgets. Christopher. D. Muir. bioRxiv 529487; doi: https://doi.org/10.1101/529487
}
