% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.AT}
\alias{f.AT}
\title{Coupled conductance photosynthesis model and energy balance model}
\usage{
f.AT(
  PFD,
  NIR = NA,
  ca,
  Tair,
  RHa,
  wind,
  precision = 0.05,
  max_it = 10,
  param,
  abso_s = 0.5,
  leaf_size = 0.04
)
}
\value{
List of different variables:

 - A: Raw assimilation of the leaf in micromol.m-2.s-1.
 
 - Ac: Rubisco limitation assimilation of the leaf in micromol.m-2.s-1.
 
 - Aj: Electron transport rate assimilation of the leaf in micromol.m-2.s-1.
 
 - Ap: TPU rate of the leaf in micromol.m-2.s-1.
 
 - Ag: Gross assimilation in micromol.m-2.s-1.
 
 - Rd: Respiration rate in micromol.m-2.s-1.
 
 - gs: Conductance of the leaf for water vapour in mol m-2 s-1.
 
 - ci: Intracellular CO2 concentration in micromol.mol-1.
 
 - ds: Leaf surface to air vapour pressure deficit in Pa.
 
 - Transp: Water transpiration in mL m-2 s-1.
 
 - gbw: boundary layer conductance in mol m-2 s-1.
 
 - cs: CO2 concentration at the leaf surface in ppm.
 
 - RHs: Relative humidity at the leaf surface (0 100).
 
 - Tleaf: Leaf temperature in K
}
\description{
Coupled conductance photosynthesis model and energy balance model
}
\details{
This function allows to calculate the photosynthesis from environmental variables PFD, RH, wind, cs and Tair.
The energy balance model is calculated using the package Tealeaves (see reference). The energy balance calculation involves the stomatal conductance and the cuticular conductance.
Here the cuticular conductance is considered to be equal to g0 as done in some TBMs even if it is probably a wrong representation.This choice was made to prevent unrealistic energy budgets when the conductance is too low (<= 0) for low light levels.
}
\examples{
leaf_physio=f.AT(PFD=seq(0,1500,50),ca=400,Tair=298,wind=2,RHa=70,param=f.make.param(g0=0.03))
plot(x=seq(0,1500,50),y=leaf_physio$A)
}
\references{
tealeaves: an R package for modelling leaf temperature using energy budgets. Christopher. D. Muir. bioRxiv 529487; doi: https://doi.org/10.1101/529487.
}
