% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canopy_scaling.R
\name{f.canopy.interception}
\alias{f.canopy.interception}
\title{Wrapper of biocro lightME and sunML function to describe the light levels inside the canopy}
\usage{
f.canopy.interception(
  meteo_hourly,
  lat,
  t.d,
  DOY,
  n_layers,
  Height,
  LAI,
  chi.l = 0.9
)
}
\arguments{
\item{meteo_hourly}{Hourly weather data frame with at least the column at (air temperature in degree C) tl (leaf temperature in degree C) rh (humidity in pc) and sr the total PAR in micro mol m-2 s-1}

\item{lat}{Latitude of the canopy to model (see lightME from biocro)}

\item{t.d}{time of the day (see lightME from biocro)}

\item{DOY}{Day of Year (see lightME from biocro)}

\item{n_layers}{Number of layers inside the canopy (max = 50, see sunML from biocro)}

\item{Height}{Total height of the canopy (see sunML from biocro)}

\item{LAI}{Total LAI of the cnaopy (see sunML from biocro)}

\item{chi.l}{Orientation of the leaves (see sunML from biocro)}
}
\value{

}
\description{
Wrapper of biocro lightME and sunML function to describe the light levels inside the canopy
}
\examples{
##Simulation of weather data
meteo_hourly=data.frame(time=0:23,rh=80,at=25,sr=sin(seq(0,pi,pi/23))*2000,tl=25)
meteo_hourly[!meteo_hourly$time\%in\%7:17,'sr']=0
##Representation of the light interception inside the canopy
canopy=f.canopy.interception(meteo_hourly=meteo_hourly,lat = 9.2801048,t.d = 0:23,DOY = 60,n_layers = 50,Height = 26,LAI = 6)
}
