% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.ds}
\alias{f.ds}
\title{Leaf water vapour pressure deficit calculation}
\usage{
f.ds(Tleaf, Tair, RH)
}
\arguments{
\item{Tleaf}{Temperature of the leaf in Kelvin}

\item{Tair}{Temperature of the air in Kelvin}

\item{RH}{Humidity of the air (0 to 100)}
}
\value{
Ds in Pascal
}
\description{
This function calculates the leaf water pressure deficit (VPDl or Ds) using the temperature of the leaf, the temperature of the air and its relative humidity
}
\examples{
f.ds(Tleaf=273.16 + 30, Tair=273.16+28, RH=70)
}
