% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.gs}
\alias{f.gs}
\title{USO model for stomatal conductance to water vapour}
\usage{
f.gs(A, cs, ds, g0, g1, Rd)
}
\arguments{
\item{A}{Raw assimilation in micromol.m-2.s-1, i-e, the assimilation in presence of respiration}

\item{cs}{CO2 at the surface of the leaf in ppm}

\item{ds}{Leaf surface to air vapour pressure deficit in Pa}

\item{g0}{Constant of the USO model, representing the conductance when A is 0, in mol.m-2.s-1}

\item{g1}{Slope parameter, between 1.14 and 3.58 KPa^0.5 (Wu et al., 2019)}

\item{Rd}{Respiration rate in micromol.m-2.s-1}
}
\value{
This function returns the optimal stomatal conductance to water vapour in mol.m-2.s-1
}
\description{
Semi-empirical model of the leaf conductance to water vapour
}
\examples{
gs=f.gs(A=30,cs=400,ds=1500,g0=0.01,g1=2,Rd=1)
}
\references{
Medlyn, B.E., Duursma, R.A., Eamus, D., Ellsworth, D.S., Colin Prentice, I., Barton, C.V.M., Crous, K.Y., de Angelis, P., Freeman, M. and Wingate, L. (2012), Reconciling the optimal and empirical approaches to modelling stomatal conductance. Glob Change Biol, 18: 3476-3476. doi:10.1111/j.1365-2486.2012.02790.x
 Wu, J, Serbin, SP, Ely, KS, et al. The response of stomatal conductance to seasonal drought in tropical forests. Glob Change Biol. 2020; 26: 823– 839. https://doi.org/10.1111/gcb.14820
}
