% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.gsmin}
\alias{f.gsmin}
\title{Calculation of the minimal conductance given by a particular coupled conductance and photosynthesis model}
\usage{
f.gsmin(
  RdRef = 0.825,
  RdHa = 46390,
  RdHd = 150650,
  RdS = 490,
  Tleaf = 300,
  cs = 400,
  ds = 1000,
  g0 = 0.02,
  g1 = 4.1,
  power = 0.5,
  model = "USO"
)
}
\value{
Minimal conductance
}
\description{
The minimal conductance of a model depends on the parameters of the model (ie g0 and g1) but also on the minimum A value, which corresponds to the dark respiration.
Knowing the minimal conductance is important because the conductance can become negative and lead to unrealistic values in photosynthesis models
}
\examples{
gs_min=f.gsmin(RdRef=	0.825,RdHa=	46390,RdHd=150650,RdS=490,Tleaf=300,cs=400,ds=1000,g0=0.02,g1=4.1,power=0.5,model="USO")
}
