% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{f.import_licor6800}
\alias{f.import_licor6800}
\title{Import Licor 6800 excel file}
\usage{
f.import_licor6800(
  file,
  column_display = c("A", "gsw", "Qin", "Ci", "Species", "Canopy", "Pheno_Age",
    "Barcode", "file")
)
}
\arguments{
\item{file}{File to import by the function}

\item{column_display}{The first lines of the file which are part of this list are displayed by this function after being imported.}
}
\value{
dataframe
}
\description{
This functions allows to import the excel file produced by LICOR as a data.frame. The files have to be open in Excel and saved before using his function so the result of the formula are calculated. The formula are sotred into the Excel file but not computed until the file is open.
}
\references{
Adapted from http://www.ericrscott.com/2018/01/17/li-cor-wrangling/
}
