% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photosynthesis_functions.R
\name{f.solv}
\alias{f.solv}
\title{Analytical solution of the coupled photosynthesis and conductance models}
\usage{
f.solv(x, y, cs, Rd, Gstar, g0, g1, power, ds, RH, model)
}
\arguments{
\item{model}{}
}
\description{
Analytical solution of the coupled photosynthesis and conductance models
}
\keyword{internal}
